change statistics to be a long

+++ WaywardBeyond.Client.Core/Skills/Listeners/LevelInfo.cs
@@ -1,3 +1,3 @@
 namespace WaywardBeyond.Client.Core.Skills.Listeners;
 
-internal record struct LevelInfo(int Level, int XP);
+internal record struct LevelInfo(int Level, long XP);
